+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.37.0] - 2025-12-10
+
+### Added
+
+- Support for `putItem` `returnValues`. This also required the introducion of the
+`PutItemResult`. When `returnValues` is set to `"NONE"` or just
+omitted, the method now returns an empty object instead of `void`. When `returnValues`
+is `ALL_OLD`, the method now returns a response of the type: `{ item: Attributes }`.
+
 ## [0.36.0] - 2025-12-08
 
 ### Added
@@ -470,6 +479,7 @@ intuitive.
 project to here.
 
 
+[0.37.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.36.0...v0.37.0
 [0.36.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.35.0...v0.36.0
 [0.35.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.34.0...v0.35.0
 [0.34.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.33.0...v0.34.0